int overflow fixed

+++ src/include/joiner.h
@@ -21,7 +21,7 @@ public:
   void addRelation(const char *file_name);
   void addRelation(Relation &&relation);
   static void appendHistogram(std::vector<std::vector<int>> histogram);
-  double estimateSelectivity(std::vector<int> histogram, int minVal, int maxVal, int bucketWidth, FilterInfo::Comparison op, uint64_t val, int nTups);
+  double estimateSelectivity(std::vector<int> histogram, uint64_t minVal, uint64_t maxVal, int bucketWidth, FilterInfo::Comparison op, uint64_t val, int nTups);
   /// Get relation
   const Relation &getRelation(unsigned relation_id);
   /// Joins a given set of relations
